requirements.txt
anyio==3.6.2
click==8.1.3
fastapi==0.95.2
h11==0.14.0
idna==3.4
numpy==1.24.3
opencv-python==4.7.0.72
pydantic==1.10.7
sniffio==1.3.0
starlette==0.27.0
typing_extensions==4.5.0
uvicorn==0.22.0
mediapipe~=0.10.0
python-dotenv~=1.0.0